order, store.let 코드 추가

+++ src/main/java/com/lch/topick/web/order/def/repository/OrderDetailRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.lch.topick.web.order.def.entity.OrderDetail;
 
-public interface OrderDetailRepository extends JpaRepository<OrderDetail, Integer>{
+public interface OrderDetailRepository extends JpaRepository<OrderDetail, Long>{
 
-	public List<OrderDetail> findByOrderListNo(Integer orderListNo);
+	public List<OrderDetail> findByOrderListNo(Long orderListNo);
 	
 } //interface
